+++ src/main/java/com/yts8/sixuniverse/member/service/MemberService.java
@@ -6,7 +6,7 @@ public interface MemberService {
 
   MemberDto findByEmail(String email);
 
-  MemberDto findById(int memberId);
+  MemberDto findById(Long memberId);
 
   void save(MemberDto memberDto);
